perf: remove `xmlns` attribute
From MDN:
The xmlns attribute is only required on the outermost svg element of SVG documents. It is unnecessary for inner svg elements or inside HTML documents.

Since it is assumed that the main case of using icons will be embedding in an HTML document, the `xmlns` attribute may be omitted by default.